requirements.txt
line-bot-sdk
flask
gunicorn
flask-ngrok
urllib3==1.26.12
requests==2.31 
python-dotenv
mongoengine
pymongo
openai

